requirements.txt
torch@https://download.pytorch.org/whl/cpu/torch-2.2.2%2Bcpu-cp38-cp38-linux_x86_64.whl
torchvision@http://download.pytorch.org/whl/cpu/torchvision-0.17.2%2Bcpu-cp38-cp38-linux_x86_64.whl
psutil~=5.9.5
pillow~=10.3.0
flask~=2.3.3
scikit-learn~=1.3.2
requests~=2.31.0
flask-cors~=4.0.0